feat(web): simplify ordinary agent creation
